requirements.txt
# Core dependencies - Python 3.12+ compatible
flask>=3.0.0
flask-cors>=4.0.0
groq>=0.4.1
cohere>=5.0,<6.0
faiss-cpu>=1.7.4,<2.0
numpy>=1.26.0
scikit-learn>=1.4.0
gunicorn>=21.2.0
python-dotenv>=1.0.0